Do more in JS
Since this is a JS course I decided to find a way to inject CSS from the JS file rather than doing the tag work in css.. kind of just an experiment. Works now. So at the moment the only file that is different between the original project files and this project is actually the JS file.